getting tripdetails and displaying it
