+++ Frontend/src/Components/map/DisplayPlaceDetails/DisplayPlaceDetails.css
